rename cmdline to _cmdline
