+++ src/stpipe/_cmdline.py
